requirements.txt
numpy
scipy
cupy-cuda12x

